# docs/api/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Agent Banking Platform API
  description: Aggregated API from all microservices
  contact:
    name: Agent Banking Platform Team
servers:
- description: Gateway
  url: http://localhost:8080
- description: Production
  url: https://api.agentbanking.example.com
- description: Sandbox
  url: https://sandbox.agentbanking.example.com
paths:
...
  /api/v1/backoffice/audit-logs:
    get:
      tags:
      - audit-log-controller (onboarding-service)
      operationId: getAuditLogs
      security:
        - bearerAuth: []
      parameters:
      - name: entityType
        in: query
        required: false
        schema:
          type: string
      - name: fromDate
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: toDate
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      required:
      - status
      - error
      type: object
      properties:
        status:
          type: string
          enum: [FAILED]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            action_code:
              type: string
              enum: [DECLINE, RETRY, REVIEW]
            trace_id:
              type: string
            timestamp:
              type: string
              format: date-time

    # ============================================================
    # TRANSACTION ORCHESTRATOR SCHEMAS (NEW - 2026-04-05)
    # ============================================================

    # Transaction Type Enum
...